Former-commit-id: 5ab5a9592552a211e3ed3eddf9ce80c238f2006f

+++ src/backend/executor/executor_context.h
@@ -44,7 +44,7 @@ class ExecutorContext {
   uint32_t num_processed = 0;
 
   ~ExecutorContext(){
-    for(size_t i=0; i < params_.GetSize(); i++){
+    for(int i=0; i < params_.GetSize(); i++){
       params_[i].FreeUninlinedData();
     }
   }
